chore: Update sign-out redirect URL
The code changes in this commit update the redirect URL after signing out. Instead of redirecting to "/admin/sign-in", the user will now be redirected to "/admin/sign-in?logout=true". This change ensures that the user is directed to the sign-in page with a query parameter indicating that they have been logged out.

```